+++ pyqgiswps/handlers/__init__.py
@@ -19,6 +19,8 @@ from .apihandler import ( # noqa F401
     JobHandler,
     ExecuteHandler,
     ResultHandler,
+    ConformanceHandler,
 )
 from .statushandler import StatusHandler                 # noqa F401
 from .storehandler import StoreHandler, DownloadHandler  # noqa F401
+from .htmlhandler import HtmlHandler # noqa F401