+++ pyqgiswps/handlers/storehandler.py
@@ -18,6 +18,11 @@ from pyqgiswps.executors.logstore import logstore
 
 from .basehandler import BaseHandler
 
+from typing import (
+    Optional,
+    Awaitable,
+)
+
 LOGGER = logging.getLogger('SRVLOG')
 
 
@@ -35,15 +40,11 @@ def _format_size( size ):
 class StoreShellMixIn():
     """ Store api implementation
     """
-    async def ls( self, uuid ):
-        """ List all files in workdir
-        """
-        path = os.path.join(self._workdir, uuid)
-        if not os.path.isdir(path):
-            raise HTTPError(404, reason="The resource does not exists")
 
-        proxy_url = self.proxy_url()
-        store_url = self.application.config['store_url']
+    # XXX DEPRECATED - will be removed at version 1.9
+    def legacy_resource_list(self, path, uuid):
+        # Legacy json schema for
+        store_url = f"{self.proxy_url()}store/{uuid}/"
 
         def file_records(rootdir, dirs, files, rootfd):
             root = Path(rootdir)
@@ -53,12 +54,47 @@ class StoreShellMixIn():
                 yield {
                     'name': name,
                     'content_length': stat.st_size,
-                    'store_url'     : store_url.format(host_url=proxy_url, uuid=uuid, file=name),
+                    'store_url'     : store_url + name,
                     'display_size'  : _format_size(stat.st_size),
                 }
-
         data = [f for args in os.fwalk(path) for f in file_records(*args)]
-        self.write_json({ "files": data })                  
+        content = { "files": data }       
+
+        return content
+
+    async def ls( self, uuid ):
+        """ List all files in workdir
+        """
+        path = os.path.join(self._workdir, uuid)
+        if not os.path.isdir(path):
+            raise HTTPError(404, reason="The resource does not exists")
+
+
+        if self._legacy:
+            # XXX DEPRECATED - will be removed at version 1.9
+            content = self.legacy_resource_list(path, uuid) 
+        else:
+            store_url = f"{self.proxy_url()}jobs/{uuid}/"
+            # Open api conformant schema
+            def file_records(rootdir, dirs, files, rootfd):
+                root = Path(rootdir)
+                for file in files:
+                    stat  = os.stat(file, dir_fd=rootfd)
+                    name  = (root / file).relative_to(path).as_posix()
+                    content_type = mimetypes.types_map.get(
+                        os.path.splitext(name)[1]
+                    ) or "application/octet-stream"
+                    yield {
+                        'href': store_url + name,
+                        'type': content_type,
+                        'name': name,
+                        'content_length': stat.st_size,
+                        'display_size'  : _format_size(stat.st_size),
+                    }
+            data = [f for args in os.fwalk(path) for f in file_records(*args)]
+            content = { "links": data }    
+
+        self.write_json(content)                  
 
     async def dnl( self, uuid, filename):
         """ Return output file from process working dir
@@ -72,15 +108,16 @@ class StoreShellMixIn():
         stat  = os.stat(full_path)
         mtime = str(stat.st_mtime)
 
-        if self.request.headers.get('If-None-Match') == mtime:
-            self.set_header('Etag', mtime)
+        etag = f"'{mtime}'"
+        if self.request.headers.get('If-None-Match') == etag:
+            self.set_header('Etag', etag)
             self.set_status(304)
             return
 
         # Set headers
         content_type = mimetypes.types_map.get(os.path.splitext(full_path)[1]) or "application/octet-stream"
         self.set_header("Content-Type", content_type)       
-        self.set_header("Etag", mtime)
+        self.set_header("Etag", etag)
 
         # Set aggresive browser caching since the resource
         # is not going to change
@@ -103,42 +140,57 @@ class StoreShellMixIn():
         # Note implemented
         raise HTTPError(501, reason="Sorry, the method is not implemented yet")  
 
+    def create_dnl_url(self, job_id: str, resource: Optional[str]=None):
+        """ Store the request and create a download url
+        """
+        token = logstore.set_json({
+            'uuid': job_id,
+            'name': resource,
+        }, self._ttl)
 
+        proxy_url = self.proxy_url()
+        now       = datetime.utcnow().timestamp()
+        self.write_json({
+            'name': resource,
+            'href': f"{proxy_url}dnl/{token}",
+            'expire_at': datetime.fromtimestamp(now+self._ttl).isoformat()+'Z',
+        })
 
 
 class StoreHandler(BaseHandler, StoreShellMixIn):
     """ Handle store requests
     """
-    def initialize(self, workdir, chunk_size=65536):
+    def initialize(self, workdir, chunk_size=65536, legacy: bool=False, ttl=30):
         super().initialize()
         self._chunk_size = chunk_size
-        self._workdir    = workdir
+        self._workdir = workdir
+        self._legacy = legacy
+        self._ttl = ttl
 
-    async def get(self, uuid, filename=None):
+    async def get(self, uuid: str, resource: Optional[str]=None) -> Awaitable[None]:
         """ Handle GET request
         """
-        if filename:
-            await self.dnl( uuid, filename)
+        if resource:
+            await self.dnl(uuid, resource)
         else:
-            await self.ls( uuid )
+            await self.ls(uuid)
 
-    async def post(self, uuid ):
-        """ Handle POST actions
+    async def put(self, uuid: str, resource: Optional[str]) -> Awaitable[None]:
+        """ Handle put actions
         """
         if not uuid:
-            raise HTTPError(403)
+            raise HTTPError(403, reason="Missing job ID")
 
         # Get the status for uuid
-        wps_status = self.application.wpsservice.get_status(uuid)
-        if wps_status is None:
+        job_status = self.application.wpsservice.get_status(uuid)
+        if job_status is None:
             raise HTTPError(404, reason="The resource does not exists") 
 
-        action = self.get_argument('ACTION') 
-        
-        if action.lower() == 'archive':
-            await self.archive(uuid)
+        command = self.get_argument('COMMAND', default="").lower()
+        if command == 'geturl':
+            self.create_dnl_url(uuid, resource)
         else:
-            raise HTTPError(400, reason="Invalid action parameter '%s'" % action)
+            raise HTTPError(400, reason=f"Invalid command parameter {command}")
 
 
 class DownloadHandler(BaseHandler, StoreShellMixIn):
@@ -147,29 +199,17 @@ class DownloadHandler(BaseHandler, StoreShellMixIn):
         1. Return a tokenized set of  url with a limited ttl
         2. Handle the url by returning the appropriate ressource
     """
-    def initialize(self, workdir, query=False,  chunk_size=65536, ttl=30):
+    def initialize(self, workdir, query=False,  chunk_size=65536):
         super().initialize()
         self._chunk_size = chunk_size
-        self._workdir    = workdir
-        self._query      = query
-        self._ttl        = ttl
+        self._workdir = workdir
 
-    def create_dnl_url(self, uuid, filename):
-        """ Store the request and create a download url
+    async def get(self, token):
+        """ Handle GET request
         """
-        token = logstore.set_json({
-            'uuid': uuid,
-            'filename': filename,
-        }, self._ttl)
-
-        proxy_url = self.proxy_url()
-        now       = datetime.utcnow().timestamp()
-        path      = "dnl/_/{}".format(token)
-        self.write_json({
-            'root': "/{}".format(path),
-            'link': "{}{}".format(proxy_url,path),
-            'expire_at': datetime.fromtimestamp(now+self._ttl).isoformat()+'Z',
-        })
+        # Download
+        uuid, resource = self.get_dnl_params(token)
+        await self.download(uuid, resource)
 
     def get_dnl_params(self, token):
         """ Retrieve the download parameters
@@ -177,19 +217,6 @@ class DownloadHandler(BaseHandler, StoreShellMixIn):
         params = logstore.get_json(token)
         if params is None:
             raise HTTPError(403)
-        return params['uuid'],params['filename']
-
-    async def get(self, uuid, resource):
-        """ Handle GET request
-        """
-        if self._query:
-            # Store the request and return 
-            # the temporary url
-            self.create_dnl_url(uuid, resource)    
-        else:
-            # Download
-            uuid,filename = self.get_dnl_params(resource)
-            await self.dnl( uuid, filename)
-
+        return params['uuid'], params['resource']